The twins drift downstream
This introduces movement. The twins move downstream at a constant rate.

This subscribes to diffs of the animation frame to keep smooth graphics
at the highest framerate.

To keep movement smooth, the position vectors are converted to floats
instead of ints. This is because at 60ish ticks a second, movement may
be under 1 px. When you throw in rounding, this leads to jerky movement
or worse, "stuck" twins when movement is rounded down to 0.

To make the math easier, this introduces a local position type with
coordinates centered on the given entity.