Add collision detection
This adds collision detection using the "axis-aligned bounding box"
algorithm. It adds a `Collidable` type that acts as a bounding box that
can be compared for collisions.

If the twins collide with a log, the game is over.

This extracts the hardcoded pixel dimensions for the twins and logs into
constants in feet. This allows us to calculate the bounding box for the
entities.