Landing on the bank loses the game
This introduces collision detection for both north and south banks. If
the player touches either then the game ends.

Since there are now multiple ways to transition to won/lost and I want
to be able to pipeline them nicely, I've added a monomorphic `andThen`
function that will only run the next check if the user is still playing.