Model infinite world as a list of sections
Previously, the world just had a hard-coded list of wolves and logs. To
make things easier to work with an infinite world (and to make it easier
to randomly generate pieces), this introduces a block-based system.

The world is divided into sections that encode for a series of blocks
that follow certain rules:

* Only one wolf "block" per section.
* At least 4 "blocks" that could contain logs in a section.
* A section can have a variable number of "blocks" that contain logs.

Sections do not store positions, simply configurations. There are 3
possible wolf configurations and 7 possible log configurations.

The Y position of logs and wolves can be derived directly from the
configuration. The X position is derived from a given "block"'s offset
in the world. All "block"s have the same width so if you know that a
given block is the `n`th from the start, you can get the X position from
the beginning of the river by `n * blockWidth`.

Wolf and log positions are re-derived every tick for both game logic in
the `update` and render logic in the `view`. This is naive and wasteful,
but good enough for now. I will revisit this if I find that it has
become a performance bottleneck.